+++ CHANGELOG.md
@@ -44,6 +44,7 @@ AQUA core complete list:
 - Added the `cdo_options: "--force"` to the definitions of the oceanic HealPix grids (#1539)
 
 AQUA diagnostic complete list:
+- ECmean: Integrating the performance indices and global mean within the `aqua_diagnostics` module (#1556)
 - Teleconnections: The `teleconnections` diagnostic is now integrated in the `aqua_diagnostics` module (#1352)
 - Teleconnections: OutputSaver for the teleconnections diagnostic (#1567, #1570)
 - Ocean3d: Fix to improve memory usage and cli (#1490)